Rename local storage helper file
