+++ src/DeVinci_frontend/components/SidebarHeader.svelte
@@ -9,7 +9,7 @@
 
   import devincilogo from "/devinci-logo.svg";
 
-  import { userHasDownloadedModel } from "../helpers/localStorage";
+  import { userHasDownloadedModel } from "../helpers/local_storage";
 
   // Reactive statement to check if the user has already downloaded at least one AI model
   $: userHasDownloadedAtLeastOneModel = userHasDownloadedModel();
