+++ src/DeVinci_frontend/helpers/user_settings.js
@@ -1,0 +1,9 @@
+import {
+  store,
+  saveChatsUserSelection
+} from "../store";
+
+let storeState;
+store.subscribe((value) => storeState = value);
+
+
