+++ src/DeVinci_frontend/pages/UserSettings.svelte
@@ -14,7 +14,7 @@
 
   import {
     syncLocalChanges
-  } from "../helpers/localStorage";
+  } from "../helpers/local_storage";
 
   let hasLoadedSettings = false;
